requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
pillow==10.2.0
pdf2image==1.17.0
paddleocr==2.10.0
paddlepaddle==2.6.2
vietocr==0.3.13
Flask==3.1.2
Werkzeug==3.1.3
boto3==1.40.23
botocore==1.40.23
confluent-kafka==2.10.0
python-dotenv==1.1.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
requests==2.32.5
